# platform/web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IntentCenter API",
    "description": "IntentCenter \u2014 DCIM, IPAM, circuits, automation, and closed-loop inventory (clean-room derived). Read-only GraphQL at /graphql.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/validation/location-type": {
      "get": {
        "tags": [
          "validation"
        ],
        "summary": "Validate Location Type",
        "operationId": "validate_location_type_v1_validation_location_type_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "`LocationType.id` in the global catalog",
              "title": "Id"
            },
            "description": "`LocationType.id` in the global catalog"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate Location Type V1 Validation Location Type Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/validation/location": {
      "get": {
        "tags": [
          "validation"
        ],
        "summary": "Validate Location",
        "operationId": "validate_location_v1_validation_location_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Org-scoped `Location.id`",
              "title": "Id"
            },
            "description": "Org-scoped `Location.id`"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate Location V1 Validation Location Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/validation/rack": {
      "get": {
        "tags": [
          "validation"
        ],
        "summary": "Validate Rack",
        "operationId": "validate_rack_v1_validation_rack_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Org-scoped `Rack.id`",
              "title": "Id"
            },
            "description": "Org-scoped `Rack.id`"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate Rack V1 Validation Rack Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/validation/device-type": {
      "get": {
        "tags": [
          "validation"
        ],
        "summary": "Validate Device Type",
        "operationId": "validate_device_type_v1_validation_device_type_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "`DeviceType.id` in the global catalog",
              "title": "Id"
            },
            "description": "`DeviceType.id` in the global catalog"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate Device Type V1 Validation Device Type Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/validation/device-role": {
      "get": {
        "tags": [
          "validation"
        ],
        "summary": "Validate Device Role",
        "operationId": "validate_device_role_v1_validation_device_role_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "`DeviceRole.id` in the global catalog",
              "title": "Id"
            },
            "description": "`DeviceRole.id` in the global catalog"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate Device Role V1 Validation Device Role Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/validation/interface": {
      "get": {
        "tags": [
          "validation"
        ],
        "summary": "Validate Interface",
        "operationId": "validate_interface_v1_validation_interface_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Interface id (must belong to a device in your org)",
              "title": "Id"
            },
            "description": "Interface id (must belong to a device in your org)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate Interface V1 Validation Interface Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/validation/cable-ends": {
      "get": {
        "tags": [
          "validation"
        ],
        "summary": "Validate Cable Endpoints",
        "description": "Check both cable endpoints in one round trip (org scope + interface existence).",
        "operationId": "validate_cable_endpoints_v1_validation_cable_ends_get",
        "parameters": [
          {
            "name": "interfaceAId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interfaceaid"
            }
          },
          {
            "name": "interfaceBId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interfacebid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate Cable Endpoints V1 Validation Cable Ends Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}